requirements.txt
opencv-python
flask
pillow
websockets
simple-websocket-server